Added sharded config generators and refactored parsing
